Decomp func_80057848 and fix func_8005D338 interface

+++ include/function_symbols.h
@@ -36,7 +36,7 @@ extern s32 func_8005C774(s32);
 extern void func_8005CA34(s32 mag, s32 time); // shake camera vertically.
 extern s32 func_8005D1B0(u16);
 extern s32 func_8005D1E8(u16);
-extern s32 func_8005D338(u16 actor_index);
+extern u8 func_8005D338(u16 actor_index);
 extern s32 func_8005D418(u16 actor_index);
 extern void func_8005DE30(void);
 extern s16 func_8005DEFC(void);
